# services/inventory/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory API
  version: 0.7.4
  description: |
    Authenticated routes use Bearer JWT; `tenant_id` is taken from the token (ARCHITECTURE §6–§7), not `X-Tenant-Id`.
    JSON errors:  via pkg/common/errorcodes. JSON success: via pkg/common/httpresponse (`success`, `data`, `meta.request_id`, list `meta.pagination`).
    Probes `/health`, `/ready`, `/api/v1/inventory/health` use `text/plain`, not §9 JSON.
    POST `/api/v1/inventory/movements/{inbound|outbound|transfer|adjustment}` require header `Idempotency-Key` (ARCHITECTURE §9); replay with the same key and identical raw JSON body returns the existing draft movement.
paths:
...
  /api/v1/inventory/dashboard/summary:
    get:
      operationId: getApiV1InventoryDashboardSummary
      tags: [dashboard]
      summary: Dashboard aggregate counts (tenant); summary is cache-aside in Redis ~30s (ARCHITECTURE §13)
      security:
        - JwtAuth: []
      responses:
        "200":
          description: Aggregate counts for the current tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummarySuccessEnvelope"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/inventory/dashboard/movements/chart:
    get:
      operationId: getApiV1InventoryDashboardMovementsChart
      tags: [dashboard]
      summary: Confirmed movement counts per UTC bucket (30 days / 12 weeks / 12 months); cache-aside ~30s (§13)
      security:
        - JwtAuth: []
      parameters:
        - name: period
          in: query
          required: false
          description: Bucket size (UTC). Daily = last 30 calendar days including today; weekly = 12 ISO weeks (Monday); monthly = 12 calendar months.
          schema:
            type: string
            enum: [daily, weekly, monthly]
            default: daily
      responses:
        "200":
          description: Time series of confirmed movements (updated_at UTC)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMovementChartSuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /api/v1/inventory/audit-logs:
    get:
      operationId: getApiV1InventoryAuditLogs
      tags: [audit]
      summary: List audit logs (tenant-scoped, §14)
      security:
        - JwtAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: entity
          in: query
          description: Filter by entity name (e.g. category, product, warehouse, movement)
          schema:
            type: string
        - name: entity_id
          in: query
          description: Filter by affected entity UUID
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          description: Filter by acting user UUID
          schema:
            type: string
            format: uuid
        - name: created_from
          in: query
          description: Inclusive lower bound (RFC3339 / ISO-8601)
          schema:
            type: string
            format: date-time
        - name: created_to
          in: query
          description: Inclusive upper bound (RFC3339 / ISO-8601)
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated audit entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/inventory/audit-logs/{auditEntity}/{auditEntityId}:
    parameters:
      - name: auditEntity
        in: path
        required: true
        schema:
          type: string
      - name: auditEntityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getApiV1InventoryAuditLogsAuditEntityAuditEntityId
      tags: [audit]
      summary: Audit history for one entity instance
      security:
        - JwtAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated audit entries for the entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
components:
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT; tenant scope from claims.
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    UnauthorizedError:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequestError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Machine-readable error message
    PaginationMeta:
      type: object
      required:
        - page
        - per_page
        - total
        - total_pages
      properties:
        page:
          type: integer
          format: int32
        per_page:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
    SuccessMeta:
      type: object
      description: §9 meta for JSON success responses (list endpoints include pagination).
      properties:
        request_id:
          type: string
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
...
    DashboardSummary:
      type: object
      required:
        - total_products
        - total_warehouses
        - movements_today
        - low_stock_count
      properties:
        total_products:
          type: integer
          format: int64
          description: Non-deleted products for the tenant
        total_warehouses:
          type: integer
          format: int64
          description: Active non-deleted warehouses
        movements_today:
          type: integer
          format: int64
          description: Movements confirmed today (UTC calendar day)
        low_stock_count:
          type: integer
          format: int64
          description: Products whose total on-hand quantity is below reorder_level
    DashboardSummarySuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/DashboardSummary"
        meta:
          $ref: "#/components/schemas/SuccessMeta"
    DashboardMovementChartPoint:
      type: object
      required: [bucket_start, movement_count]
      properties:
        bucket_start:
          type: string
          format: date
          description: Start of bucket in UTC (calendar day, ISO week Monday, or month start)
        movement_count:
          type: integer
          format: int64
          description: Confirmed movements with updated_at in this bucket
    DashboardMovementChart:
      type: object
      required: [period, points]
      properties:
        period:
          type: string
          enum: [daily, weekly, monthly]
        points:
          type: array
          items:
            $ref: "#/components/schemas/DashboardMovementChartPoint"
    DashboardMovementChartSuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: "#/components/schemas/DashboardMovementChart"
        meta:
          $ref: "#/components/schemas/SuccessMeta"
...
    AuditLog:
      type: object
      required:
        - id
        - tenant_id
        - action
        - entity
        - entity_id
        - created_at
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        action:
          type: string
        entity:
          type: string
        entity_id:
          type: string
          format: uuid
        before_data:
          type: object
          nullable: true
          additionalProperties: true
        after_data:
          type: object
          nullable: true
          additionalProperties: true
        ip_address:
          type: string
          nullable: true
        user_agent:
          type: string
          nullable: true
        request_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    AuditLogListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        meta:
          $ref: "#/components/schemas/SuccessMeta"